+++ lib/gen/flutterblue.pbjson.dart
@@ -106,6 +106,13 @@ const ScanSettings$json = const {
       '10': 'androidScanMode'
     },
     const {'1': 'service_uuids', '3': 2, '4': 3, '5': 9, '10': 'serviceUuids'},
+    const {
+      '1': 'allow_duplicates',
+      '3': 3,
+      '4': 1,
+      '5': 8,
+      '10': 'allowDuplicates'
+    },
   ],
 };