+++ lib/src/flutter_blue.dart
@@ -19,9 +19,9 @@ class FlutterBlue {
       return;
     });
 
-    // Send the log level to the underlying platforms.
-    setLogLevel(logLevel);
+    _setLogLevelIfAvailable();
   }
+
   static FlutterBlue _instance = new FlutterBlue._();
   static FlutterBlue get instance => _instance;
 
@@ -66,6 +66,13 @@ class FlutterBlue {
         .then((p) => p.map((d) => BluetoothDevice.fromProto(d)).toList());
   }
 
+  _setLogLevelIfAvailable() async {
+    if (await isAvailable) {
+      // Send the log level to the underlying platforms.
+      setLogLevel(logLevel);
+    }
+  }
+
   /// Starts a scan for Bluetooth Low Energy devices
   /// Timeout closes the stream after a specified [Duration]
   Stream<ScanResult> scan({
@@ -73,9 +80,11 @@ class FlutterBlue {
     List<Guid> withServices = const [],
     List<Guid> withDevices = const [],
     Duration timeout,
+    bool allowDuplicates = false,
   }) async* {
     var settings = protos.ScanSettings.create()
       ..androidScanMode = scanMode.value
+      ..allowDuplicates = allowDuplicates
       ..serviceUuids.addAll(withServices.map((g) => g.toString()).toList());
 
     if (_isScanning.value == true) {
@@ -128,12 +137,14 @@ class FlutterBlue {
     List<Guid> withServices = const [],
     List<Guid> withDevices = const [],
     Duration timeout,
+    bool allowDuplicates = false,
   }) async {
     await scan(
             scanMode: scanMode,
             withServices: withServices,
             withDevices: withDevices,
-            timeout: timeout)
+            timeout: timeout,
+            allowDuplicates: allowDuplicates)
         .drain();
     return _scanResults.value;
   }